Refactor the Attribute class and make it conform with the comparable interface
